backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
pydantic
python-dotenv
chromadb
langchain
langchain-community
langchain-groq
langchain-google-genai
ollama
langchain-huggingface
langchain-chroma
sentence-transformers
docx2txt
pypdf
python-docx
fpdf2
openpyxl
python-pptx
bcrypt
python-jose[cryptography]
slowapi
cryptography
